requirements.txt
python-can==4.3.1
numpy==1.26.4
scipy==1.11.4
streamlit==1.36.0
scikit-learn==1.3.2
